Add bookAction and modify bookreducer
