+++ src/redux/book/reducer/bookReducer.js
@@ -1,0 +1,14 @@
+const initialState = [];
+
+const bookReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case 'ADD_BOOK':
+      return [...state.action.payLoad];
+    case 'REMOVE_BOOK':
+      return state.filter((item) => item.id !== action.payLoad);
+    default:
+      return state;
+  }
+};
+
+export default bookReducer;
